+++ tests/integration/dynamic/python/__main__.py
@@ -3,7 +3,6 @@
 from pulumi import ComponentResource
 from pulumi.dynamic import Resource, ResourceProvider
 
-
 class MyResourceProvider(ResourceProvider):
     def create(self, props):
         return {'id': "deadbeef", 'outs': props}